requirements.txt
flask
grpcio 
googleapis-common-protos 
redis